requirements.txt
PyYAML
paho-mqtt